requirements.txt
jaro-winkler==2.0.3
sqlescapy==1.0.1
pysftp==0.2.9
slack-sdk==3.26.2
zipfile36==0.1.3
apache-airflow-providers-microsoft-mssql==3.3.2
apache-airflow-providers-postgres==5.4.0
apache-airflow-providers-mysql==5.0.0
apache-airflow-providers-ssh==2.3.0
chardet>=3.0.4
pyarrow==12.0.1
# pyspark==3.4.4  # Uncomment to enable Spark — see README "Optional: Spark Cluster"
